+++ tests/dtale/test_views.py
@@ -160,6 +160,7 @@ def test_startup(unittest):
                 "unique_ct": 2,
                 "kurt": "nan",
                 "skew": "nan",
+                "coord": None,
             },
             next(
                 (
@@ -1336,6 +1337,7 @@ def test_get_data(unittest, test_data):
                         hasMissing=0,
                         hasOutliers=0,
                         unique_ct=1,
+                        skew=0,
                     ),
                     dict(
                         dtype="int64",
@@ -1351,6 +1353,7 @@ def test_get_data(unittest, test_data):
                         unique_ct=50,
                         kurt=-1.2,
                         skew=0,
+                        coord=None,
                     ),
                     dict(
                         dtype="int64",
@@ -1366,6 +1369,7 @@ def test_get_data(unittest, test_data):
                         unique_ct=1,
                         kurt=0,
                         skew=0,
+                        coord=None,
                     ),
                     dict(
                         dtype="float64",
@@ -1381,6 +1385,7 @@ def test_get_data(unittest, test_data):
                         unique_ct=1,
                         kurt=0,
                         skew=0,
+                        coord=None,
                     ),
                     dict(
                         dtype="string",
@@ -1617,6 +1622,7 @@ def test_get_data(unittest, test_data):
                     unique_ct=1,
                     kurt=0,
                     skew=0,
+                    coord=None,
                 ),
                 "should update dtypes on data structure change",
             )
@@ -1926,6 +1932,23 @@ def test_get_column_analysis_kde():
             )
             response_data = json.loads(response.data)
             assert len(response_data["kde"]) == 51
+
+
+@pytest.mark.unit
+def test_get_column_analysis_geolocation(unittest):
+    df = pd.DataFrame(dict(a=[1, 2, 3], b=[3, 4, 5]))
+    with app.test_client() as c:
+        with ExitStack() as stack:
+            stack.enter_context(mock.patch("dtale.global_state.DATA", {c.port: df}))
+            settings = {c.port: {}}
+            stack.enter_context(mock.patch("dtale.global_state.SETTINGS", settings))
+            response = c.get(
+                "/dtale/column-analysis/{}".format(c.port),
+                query_string=dict(col="a", type="geolocation", latCol="a", lonCol="b"),
+            )
+            response_data = json.loads(response.data)
+            unittest.assertEqual(response_data["lat"], [1, 2, 3])
+            unittest.assertEqual(response_data["lon"], [3, 4, 5])
 
 
 CORRELATIONS_CODE = """# DISCLAIMER: 'df' refers to the data you passed in when calling 'dtale.show'
@@ -2972,6 +2995,7 @@ def test_200():
         "/dtale/static/images/projections/miller.png",
         "/dtale/static/images/map_type/choropleth.png",
         "/dtale/static/maps/usa_110m.json",
+        "/dtale/network/{port}",
     ]
     with app.test_client() as c:
         with ExitStack() as stack:
@@ -3547,7 +3571,7 @@ def test_update_theme():
             c.get("/dtale/update-theme", query_string={"theme": "dark"})
             assert app_settings["theme"]
             response = c.get("/dtale/main/{}".format(c.port))
-            assert '<body class="dark-mode">' in str(response.data)
+            assert '<body class="dark-mode"' in str(response.data)
 
 
 @pytest.mark.unit
@@ -3613,3 +3637,23 @@ def test_build_row_text():
                 data={"rows": json.dumps([1]), "columns": json.dumps(["a"])},
             )
             assert resp.data == b"2\n"
+
+
+@pytest.mark.unit
+def test_sorted_sequential_diffs():
+    import dtale.views as views
+
+    df, _ = views.format_data(pd.DataFrame(dict(a=[1, 2, 3, 4, 5, 6])))
+    with build_app(url=URL).test_client() as c:
+        with ExitStack() as stack:
+            stack.enter_context(mock.patch("dtale.global_state.DATA", {c.port: df}))
+
+            resp = c.get("/dtale/sorted-sequential-diffs/{}/a/ASC".format(c.port))
+            data = resp.json
+            assert data["min"] == "1"
+            assert data["max"] == "1"
+
+            resp = c.get("/dtale/sorted-sequential-diffs/{}/a/DESC".format(c.port))
+            data = resp.json
+            assert data["min"] == "-1"
+            assert data["max"] == "-1"